Cleanup.
* Fix the Magic-Q for u16
* strange_words: fix the rotations
