+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-    ".",
     "crates/*",
 ]
 
@@ -13,14 +12,4 @@ typenum = "^1"
 zeroize = "^1"
 
 rc5-core = {path = "crates/rc5-core"}
-rc5-original-api = {path = "crates/rc5-original-api"}
 
-
-[package]
-name = "rc5_test"
-version = "0.1.0"
-authors = ["Ben Marsh <ben.marsh@mintlayer.org>"]
-edition = "2018"
-
-[dependencies]
-rc5-original-api = {workspace = true}
