+++ crates/rc5-core/src/algs/key_table.rs
@@ -8,7 +8,7 @@ use crate::traits::{
     WordBytes, WordSize,
 };
 
-pub fn key_bytes_to_words<Word, B, WBR>(key: &KeyBytes<B>, l_table: &mut KeyLTable<B, Word>)
+pub fn l_table_init<Word, B, WBR>(key: &KeyBytes<B>, l_table: &mut KeyLTable<B, Word>)
 where
     B: Add<Word::ByteLen>,
     Sum<B, Word::ByteLen>: Sub<typenum::U1>,
@@ -19,8 +19,6 @@ where
     B: ArrayLength<u8>,
     Word: WordSize,
     WBR: WordByteRepr<Word>,
-
-    Word: core::fmt::Debug,
 {
     let key_bytes = key.as_slice();
     let key_words = l_table.as_mut_slice();
@@ -36,10 +34,6 @@ where
 
         *dst_word = WBR::from_bytes(&w_bytes);
     }
-
-    // for (idx, word) in l_table.iter().enumerate() {
-    //     eprintln!("L[{}]: {:06X?}", idx, word);
-    // }
 }
 
 pub fn s_table_init<Word, R, M, A>(s_table: &mut KeySTable<R, Word>)
@@ -49,17 +43,11 @@ where
     KeySTableSize<R>: ArrayLength<Word>,
     Sum<R, typenum::U1>: Mul<typenum::U2>,
     R: Add<typenum::U1>,
-
-    Word: core::fmt::Debug,
 {
     s_table[0] = M::P;
     for i in 1..KeySTableSize::<R>::USIZE {
         s_table[i] = A::add(&s_table[i - 1], &M::Q);
     }
-
-    // for (idx, word) in s_table.iter().enumerate() {
-    //     eprintln!("S[{}]: {:06X?}", idx, word);
-    // }
 }
 
 pub fn s_table_mix_secret_key<Word, B, R, A, M>(
@@ -80,8 +68,6 @@ pub fn s_table_mix_secret_key<Word, B, R, A, M>(
     Word: WordSize + Default + Copy,
     A: Arithmetics<Word>,
     M: Magic<Word>,
-
-    Word: core::fmt::Debug,
 {
     let mut i: usize = 0;
     let mut j: usize = 0;
@@ -93,11 +79,9 @@ pub fn s_table_mix_secret_key<Word, B, R, A, M>(
     for _ in 0..(3 * t.max(c)) {
         a = A::rotl(&A::add(&s_table[i], &A::add(&a, &b)), &M::THREE);
         s_table[i] = a;
-        // eprintln!("S[{}]: {:06X?}", i, a);
 
         b = A::rotl(&A::add(&l_table[j], &A::add(&a, &b)), &A::add(&a, &b));
         l_table[j] = b;
-        // eprintln!("L[{}]: {:06X?}", j, b);
 
         i = (i + 1) % t;
         j = (j + 1) % c;
@@ -116,7 +100,7 @@ fn test_key_bytes_to_words_01() {
 
     let mut key_words = KeyLTable::<KS, W>::default();
 
-    key_bytes_to_words::<W, KS, LittleEndian>(key_bytes, &mut key_words);
+    l_table_init::<W, KS, LittleEndian>(key_bytes, &mut key_words);
 
     assert_eq!(key_words.as_slice(), &[0x01010101u32; 4]);
 }
