+++ crates/rc5-rustcrypto-api/src/tests/krovetz18.rs
@@ -14,7 +14,6 @@ fn rc5_8_12_4() {
 }
 
 #[test]
-#[ignore]
 fn rc5_16_16_8() {
     run_case::<RC5<u16, U16, U8>>("0001020304050607", "00010203", "23A8D72E")
 }
