+++ crates/rc5_test/Cargo.toml
@@ -1,10 +1,13 @@
 [package]
-name = "rc5-original-api"
+name = "rc5_test"
 version = "0.1.0"
-edition = "2021"
+authors = ["Ben Marsh <ben.marsh@mintlayer.org>"]
+edition = "2018"
+
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rc5-core = {workspace = true}
 typenum = {workspace = true}
+
+rc5-core = {workspace = true}
